requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
Faker==18.11.2
pydantic==2.4.2
pydantic_core==2.10.1
pytest==7.4.2
requests==2.31.0
selenium==4.14.0
